fix wrong data type for loss function

+++ train.py
@@ -127,7 +127,7 @@ def _eval(model, dataloader, c_matrix=None):
     total = 0
     for test_data in dataloader:
         X, y = test_data
-        X, y = X.cuda(), y.long().cuda()
+        X, y = X.cuda(), y.float().cuda()
 
         y_pred = model(X)
         total += y.size(0)
